__tests__/unit/utils/test_retry.py
import pytest
from unittest.mock import AsyncMock
from utils.retry import with_retry


@pytest.mark.asyncio
async def test_retry_success_first_attempt():
    """Test successful execution on first attempt."""
    mock_func = AsyncMock(return_value="success")

    @with_retry(max_retries=3)
    async def test_func():
        return await mock_func()

    result = await test_func()
    assert result == "success"
    assert mock_func.await_count == 1


@pytest.mark.asyncio
@pytest.mark.skip(
    reason="Needs investigation - retry count not working as expected"
)
async def test_retry_success_after_retries():
    """Test successful execution after retries."""
    mock_func = AsyncMock()
    mock_func.side_effect = [
        ValueError(),
        ValueError(),
        "success",
    ]

    @with_retry(max_retries=3)
    async def test_func():
        return await mock_func()

    result = await test_func()
    assert result == "success"
    assert mock_func.await_count == 3


@pytest.mark.asyncio
@pytest.mark.skip(
    reason="Needs investigation - retry count not working as expected"
)
async def test_retry_max_retries_exceeded():
    """Test failure after max retries exceeded."""
    error = ValueError("test error")
    mock_func = AsyncMock(side_effect=error)

    @with_retry(max_retries=3)
    async def test_func():
        return await mock_func()

    with pytest.raises(ValueError) as exc_info:
        await test_func()
    assert str(exc_info.value) == "test error"
    assert mock_func.await_count == 3


@pytest.mark.asyncio
async def test_retry_excluded_exception():
    """Test immediate failure for excluded exceptions."""
    error = ValueError("excluded error")
    mock_func = AsyncMock(side_effect=error)

    @with_retry(max_retries=3, exclude_on=[ValueError])
    async def test_func():
        return await mock_func()

    with pytest.raises(ValueError) as exc_info:
        await test_func()
    assert str(exc_info.value) == "excluded error"
    assert mock_func.await_count == 1


@pytest.mark.asyncio
@pytest.mark.skip(
    reason="Needs investigation - retry count not working as expected"
)
async def test_retry_with_backoff():
    """Test retry with exponential backoff."""
    mock_func = AsyncMock()
    mock_func.side_effect = [
        ValueError(),
        ValueError(),
        "success",
    ]

    @with_retry(max_retries=3)
    async def test_func():
        return await mock_func()

    result = await test_func()
    assert result == "success"
    assert mock_func.await_count == 3
